Fixed an issue with the model

+++ src/livro_models/repository/LivroRepository.js
@@ -1,7 +1,8 @@
 import Livro from "../schemas/Livro.js";
+
 export class LivroRepository {
 	static getAllLivros = async () => {
-		return Livro.find().populate("autor");
+		return Livro.find().populate("author");
 	};
 
 	static getLivroById = async id => {
@@ -14,6 +15,16 @@ export class LivroRepository {
 	};
 
 	static deleteLivro = async id => {
+		const livro = await Livro.findById(id).populate("author");
+
+		if (!livro) {
+			throw new Error("Livro not found");
+		}
+
+		if (livro.author.livrosCount > 0) {
+			throw new Error("Cannot delete livro. Author has associated books.");
+		}
+
 		return Livro.findByIdAndDelete(id);
 	};
